+++ sts/TigerWhale/src/main/webapp/WEB-INF/mypage/mypage-email.jsp
@@ -2,6 +2,7 @@
     pageEncoding="UTF-8"%>
 <!DOCTYPE html>
 <html lang="en">
+<!--  
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -11,6 +12,8 @@
     <title>Document</title>
 
 </head>
+-->
+
 <body>
     <div class="container">
         <div class="row">
