+++ sts/TigerWhale/src/main/webapp/WEB-INF/mypage/mypage-name.jsp
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
+<!--  
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -9,6 +10,7 @@
     <title>Document</title>
 
 </head>
+-->
 <body>
     <div class="container">
         <div class="row">
