+++ sts/TigerWhale/src/main/webapp/WEB-INF/tiles/templateMypage.jsp
@@ -8,10 +8,15 @@ detailBoard/detailBuy<%@ page language="java" contentType="text/html; charset=UT
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/bootstrap/css/bootstrap.css">
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/bootstrap/css/bootstrap-theme.css">
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/bootstrap/css/bootstrap.min.css">
-<!-- 부트스트랩 js참조 -->
 <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.3/css/all.css" integrity="sha384-SZXxX4whJ79/gErwcOYf+zWLeJdY/qpuqC4cAa9rOGUstPomtqpuNWT9wdPEn2fk" crossorigin="anonymous">
 <script src="${pageContext.request.contextPath}/resources/bootstrap/js/bootstrap.js"></script>
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/header.css ">
+
+<!-- //////////////// -->
+<!-- /////mypage///// -->
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/mypage.CSS">
+
 
 </head>
 <body>
