+++ sts/TigerWhale/src/main/webapp/WEB-INF/tiles/tiles.xml
@@ -7,15 +7,15 @@
 <tiles-definitions>
 		<!-- 1. 타일즈 파일이 가져야하는 레이아웃형식을 정의 -->
 	<!-- name속성은 상속받아서 사용할 이름, template은 타일즈템플릿의 경로 -->
-	<definition name="templateHeader" template="/WEB-INF/tiles/templateheader.jsp">
+	<definition name="templateMypage" template="/WEB-INF/tiles/templateMypage.jsp">
 		<put-attribute name="header" value="/WEB-INF/tiles/header.jsp" />
 		<put-attribute name="body" value="" />
 	</definition>
 	
 	<!-- 위에 선언된 타일즈템플릿을 상속 -->
 	<!-- name속성은 controller에서 연결할 이름 -->
-	<definition name="tiles/*" extends="templateHeader">
-		<put-attribute name="body" value="/WEB-INF/tiles/{1}.jsp" />
+	<definition name="mypage/*" extends="templateMypage">
+		<put-attribute name="body" value="/WEB-INF/mypage/{1}.jsp" />
 	</definition>
